+++ src/lib-public/FA1-generator.ts
@@ -21,6 +21,7 @@ import { Faktura } from './types/fa1.types';
 import { ZamowienieKorekta } from './enums/invoice.enums';
 import { AdditionalDataTypes } from './types/common.types';
 import { getSimplifiedPageSize, SIMPLIFIED_PAGE_MARGINS } from './utils/simplified-page-size';
+import { Position } from '../shared/enums/common.enum';
 
 pdfMake.vfs = pdfFonts.vfs;
 
@@ -54,6 +55,13 @@ export function generateFA1(invoice: Faktura, additionalData: AdditionalDataType
     ];
   const docDefinition: TDocumentDefinitions = {
     content,
+    footer: (currentPage, pageCount) => {
+      return {
+        text: currentPage.toString() + ' z ' + pageCount,
+        alignment: Position.RIGHT,
+        margin: [0, 0, 40, 0],
+      };
+    },
     ...(additionalData?.simplifiedMode
       ? {
         pageSize: getSimplifiedPageSize(additionalData),