+++ src/shared/consts/FA.const.ts
@@ -1,0 +1,5 @@
+export {
+  FormaPlatnosci,
+  TAXPAYER_STATUS,
+  TRodzajFaktury,
+} from './const';